+++ src/main/java/net/mcreator/spasm/item/PebbleItem.java
@@ -1,0 +1,9 @@
+package net.mcreator.spasm.item;
+
+import net.minecraft.world.item.Item;
+
+public class PebbleItem extends Item {
+	public PebbleItem(Item.Properties properties) {
+		super(properties);
+	}
+}