+++ src/main/java/net/mcreator/spasm/procedures/DirtDropPebbleProcedure.java
@@ -1,0 +1,45 @@
+package net.mcreator.spasm.procedures;
+
+import net.neoforged.neoforge.event.level.block.BreakBlockEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.bus.api.Event;
+
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.resources.Identifier;
+import net.minecraft.core.BlockPos;
+
+import net.mcreator.spasm.init.SpasmModItems;
+
+import javax.annotation.Nullable;
+
+@EventBusSubscriber
+public class DirtDropPebbleProcedure {
+	@SubscribeEvent
+	public static void onBlockBreak(BreakBlockEvent event) {
+		execute(event, event.getLevel(), event.getPos().getX(), event.getPos().getY(), event.getPos().getZ(), event.getState());
+	}
+
+	public static void execute(LevelAccessor world, double x, double y, double z, BlockState blockstate) {
+		execute(null, world, x, y, z, blockstate);
+	}
+
+	private static void execute(@Nullable Event event, LevelAccessor world, double x, double y, double z, BlockState blockstate) {
+		if (blockstate == Blocks.GRASS_BLOCK.defaultBlockState() || blockstate == Blocks.DIRT_PATH.defaultBlockState() || blockstate.is(BlockTags.create(Identifier.parse("minecraft:dirt")))) {
+			world.destroyBlock(BlockPos.containing(x, y, z), false);
+			if (Math.random() <= 0.005) {
+				if (world instanceof ServerLevel _level) {
+					ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(SpasmModItems.PEBBLE.get()));
+					entityToSpawn.setPickUpDelay(10);
+					_level.addFreshEntity(entityToSpawn);
+				}
+			}
+		}
+	}
+}