+++ releases/CHANGELOG.md
@@ -1,1 +1,9 @@
-- Stripped Dirt can now craft into 4 Dirt Clumps
+- Added Cobblestone Shovel
+    - Same as Wooden Dirt Shovel except:
+        - More durability
+        - 2 in enchantability
+
+- Dirt has a random change of dropping a pebble
+
+- Added Pebble
+    - 9 craft into a block of cobblestone